feat(musicplayer): add pkill, song info
Previously, the PID of the process was being used to stop the process
when a new song is being played. Now I use pkill to stop any
powershell.exe processes which should take care of stopping all audio.

Also added the song artist and title info (outputted to a log file
`now-playing.txt`) for tmux to use.