feat(tmux): change trigger key, edit status
* I changed tmux's trigger to be <C-s> instead of <C-a> since <C-a> was
a Vim command I use more frequently than I thought (to increment
numbers).
* I also therefore, changed the bindings for switching between tmux
panes. The mod key was, for some reason, the <ESC> key and that was no
bueno. So I changed the keymaps to be just the tmux trigger and then vim
directional bindings.
* I am also experimenting with my `wsl-musicplayer` now, and thus
changed my music info script to update my tmux status to revert back to
calling a powershell script. I am going to work more on this. So it
will probably change in the future.