chore(neovim): change markdown remaps
I had to change the <CR> remap to follow links for now since the end of
various remaps being <CR> was causing issues in markdown files. I have
changed it to <leader>g for GO.

A couple other remaps I added were for going back with <BS> which just
mimics <C-o> for now and <leader>a to trigger Tabularize's align
function on `|`s.