chore(polybar/volume): add new volume module
this replaces the pulseaudio module with a custom volume module. there
is a polybar ipc hook that will update the module using the `volume`
script. the `volume` script has been updated to use `pactl` instead of
ponymix for more fine grained control of output. The main problem I was
having was that for certain sinks (my THR amp via bluetooth), the sink
volume control was not working and I had to manually update the
sink-input volume via pavucontrol. This updates `volume` to synchronize
the volume control over the current active sink as well as **ALL** sink
inputs. Basically, all application volumes will be tied to the current
active sink volume.