chore(iterm): completely rehaul color/font changing script
using the internal colorscheme option instead of directly updating the
dynamic profile. this is much faster but the only caveat is that, when a
new colorscheme file is generated, it needs to be manually imported into
iTerm via the options menus. existing colorscheme files (already
imported by this method the first time), don't need any additional
manual tasks.