chore(neovim): increase foldlevel to 20
I never hit this foldlevel before today when I was working with a yaml
code block embedded inside a markdown doc. The yaml had more than 10
levels of folding I believe and I was confused as to why the initial
fold on page load was concealing this yaml partly. I think it is
reasonable to not have text folded by default for myself and while I can
try increasing this to a high value later, 20 will suffice for now.