feat(neovim): add interactive shell to load shell aliases into vim
this enables vim to be loaded with an interactive shell and pulling all
aliases defined in zsh to hopefully work. I have only tested this with
the open -> xdg-open aliases used in pandoc markdown render function
`<leader>pp` and this seems to work. I should keep an eye whether this
breaks any other functionality.