feat(neovim): increase width cutoff for statusline
the statusline can get crowded when the window size is small enough.
There is a cutoff of character blocks that will transition to a less
dense version of the statusline. Earlier, this used to be 100
characters, i.e. for less than 100 character blocks for the window, the
statusline would show a less dense version. This proved to be
insufficient and has now been increased to 150