feat(choose): add choose launcher + scripts
https://github.com/chipsenkbeil/choose

choose is a rofi/dmenu like launcher for macos that takes in input from
stdin and spits it out to stdout. I can use this to run script options
similar to rofi in linux. currently I have a script called bookmarks to
choose my browser bookmarks irrespective of the browser program I use.

there's also modifications to the colorscheme script to incorporate
certain colors to the choose GUI.