feat(sketchybar): improve jira script functionality
This commit enhances the jira script in the sketchybar configuration. It
changes the grep pattern to "t rex" to go with the new team name for my
jira and adds new sketchybar commands to provide feedback during the
script execution. The error message is now more informative, including
the values of SPRINT and END. The condition for setting the sketchybar
label is removed, so the label is always updated.