feat(zsh): enhance git aliases and functions
This commit enhances the git aliases and functions in the zsh
configuration. It replaces the 'gpo' and 'gpush' aliases with
'fuzzyGitPullOrigin' and 'fuzzyGitPushOrigin' respectively. It also
introduces a new function 'showFuzzyGitBranches' to display a list of
git branches using fzf. The 'fuzzyGitCheckout' function is updated to
use 'showFuzzyGitBranches'. The 'fuzzyGitPullOrigin' and
'fuzzyGitPushOrigin' functions are added to pull and push to a branch
selected using fzf.