refactor(pandoc): streamline CSS and update markdown
This commit refactors the CSS for the pandoc configuration, removing
redundant styles and adding new ones for code blocks and navigation.
The font for code blocks is changed to 'LigaSauceCodePro' and the
navigation is now evenly spaced. The hover effect for links is also
updated. In the markdown file, new examples for checkboxes in lists are
added. The HTML template is simplified, removing many inline styles.