refactor(neovim): update theme highlights and gitsigns
This commit refactors the theme highlights and gitsigns in the Neovim
configuration. It changes the 'darkeraccent' to be lightened based on
the current mode. It also updates the highlights for various elements
like variables, fields, and punctuation. The markdown title highlights
are changed to use the 'markup.heading' group. The gitsigns are updated
to include staged changes and use darker colors for them. The
'GitSignsChangeDelete' highlight is changed to not have a background
color.