refactor(neovim): update cmp plugin configuration
This commit refactors the cmp plugin configuration in the Neovim setup.
It removes the TODO comment about lazy loading and the unnecessary
newline. It adds the mapping preset for the cmdline setup. The key
bindings for selecting the previous and next items are changed to
"<C-p>" and "<C-n>", respectively and not overload the tab key. The
complex mappings for "<Tab>" and "<S-Tab>" are removed. The buffer
source is updated to include all buffers in the current windows. The
kind icons formatting is simplified.