refactor(neovim): simplify gitsigns configuration
This commit simplifies the gitsigns configuration in the Neovim setup.
It removes the individual highlight groups for each sign and uses the
default ones. The text for the delete sign is changed to '_', the
topdelete sign to '‾', and the untracked sign to '┆'. It also adds
staged signs with the same text. The 'signs_staged_enable' option is
set to true, and the 'auto_attach' option is added and set to true. The
'yadm' option is removed as it is not used.