feat(skhd/yabai): add floating terminal shortcut and rule
This commit adds a new shortcut in skhd configuration to open a floating
terminal with Neovim opened on the init.lua file. It also adds a new
rule in yabai configuration to make windows with the title "Float" float
above other windows, be sticky, and have a specific grid layout. This
enhancement provides a quick way to access and edit the Neovim
configuration.