feat(zsh): macos enable git, plugins and add pylsp to LSP servers
This commit enables git and zsh plugins in the zsh configuration for
macOS. It also adds 'pylsp' to the list of Language Server Protocol
(LSP) servers to be installed in vim. This change enhances the
development experience by providing git integration, additional zsh
features, and Python language support in vim.